Some small fixes, details within
Removed the use of touchforeignPointer, no longer needed when passing around the input ByteString. Using <$> instead of `fmap` because ugly. removed use of getDestProgress in decodeSequence, realised that the results are meaningless if a new ByteString is fed in half way through the computation; decodeSequence now returns the sum of the literals length and the match length (if any).